Change location of the test applications. Rename files
