+++ apps/tpg_algorithms_emulator_WIB2.cxx
@@ -7,7 +7,7 @@
  */
 
 // DUNE-DAQ
-#include "fddetdataformats/WIBEthFrame.hpp"
+#include "fddetdataformats/WIB2Frame.hpp"
 #include "fddetdataformats/WIBEthFrame.hpp"
 
 #include "iomanager/IOManager.hpp"
@@ -15,13 +15,13 @@
 #include "readoutlibs/utils/RateLimiter.hpp"
 #include "readoutlibs/utils/FileSourceBuffer.hpp"
 #include "readoutlibs/utils/BufferedFileWriter.hpp"
-#include "fdreadoutlibs/DUNEWIBEthTypeAdapter.hpp"
-#include "fdreadoutlibs/wibeth/WIBEthFrameProcessor.hpp"
+#include "fdreadoutlibs/DUNEWIBSuperChunkTypeAdapter.hpp"
+#include "fdreadoutlibs/wib2/WIB2FrameProcessor.hpp"
 
-#include "fdreadoutlibs/wibeth/tpg/ProcessAVX2.hpp"
-#include "fdreadoutlibs/wibeth/tpg/ProcessRSAVX2.hpp"
-#include "fdreadoutlibs/wibeth/tpg/ProcessNaive.hpp"
-#include "fdreadoutlibs/wibeth/tpg/ProcessNaiveRS.hpp"
+#include "fdreadoutlibs/wib2/tpg/ProcessAVX2.hpp"
+#include "fdreadoutlibs/wib2/tpg/ProcessRSAVX2.hpp"
+#include "fdreadoutlibs/wib2/tpg/ProcessNaive.hpp"
+#include "fdreadoutlibs/wib2/tpg/ProcessNaiveRS.hpp"
 
 
 #include "readoutlibs/models/DefaultRequestHandlerModel.hpp"
@@ -58,14 +58,13 @@ struct swtpg_output{
   uint64_t timestamp;
 };
 
-int WIBEth_FRAME_SIZE = 7200;
-int TIME_TEST = 120;
+
 unsigned int total_hits = 0;
 bool first_hit = true;
 
-dunedaq::fdreadoutlibs::WIBEthFrameHandler fh;
+dunedaq::fdreadoutlibs::WIB2FrameHandler fh(0);
 
-std::function<void(swtpg_wibeth::ProcessingInfo<swtpg_wibeth::NUM_REGISTERS_PER_FRAME>& info)> m_assigned_tpg_algorithm_function;
+std::function<void(swtpg_wib2::ProcessingInfo<swtpg_wib2::NUM_REGISTERS_PER_FRAME>& info, size_t channel_offset)> m_assigned_tpg_algorithm_function;
 
 std::string select_algorithm = "";
 std::string select_implementation = "";
@@ -109,7 +108,7 @@ void save_hit_data( triggeralgs::TriggerPrimitive trigprim, std::string source_n
 
 
 // Function to save raw ADC data to a file (only for debugging) 
-void save_raw_data(swtpg_wibeth::MessageRegisters register_array, 
+void save_raw_data(swtpg_wib2::MessageRegisters register_array, 
 	       uint64_t t0, int channel_number,
            std::string source_name)
 {
@@ -133,28 +132,28 @@ void save_raw_data(swtpg_wibeth::MessageRegisters register_array,
   uint64_t t_current= t0 ; 
   
   const uint16_t* input16 = register_array.data();
-  for (auto ichan = 0; ichan < swtpg_wibeth::NUM_REGISTERS_PER_FRAME * swtpg_wibeth::SAMPLES_PER_REGISTER; ++ichan) {
-    const size_t register_index = ichan / swtpg_wibeth::SAMPLES_PER_REGISTER;
-    if (register_index >= swtpg_wibeth::NUM_REGISTERS_PER_FRAME)
+  for (size_t ichan = 0; ichan < swtpg_wib2::NUM_REGISTERS_PER_FRAME * swtpg_wib2::SAMPLES_PER_REGISTER; ++ichan) {
+    const int register_index = ichan / swtpg_wib2::SAMPLES_PER_REGISTER;
+    if (register_index < 0 || register_index >= swtpg_wib2::NUM_REGISTERS_PER_FRAME)
        continue;
 
     // Parse only selected channel number. To select all channels choose -1
     if (ichan == channel_number || channel_number == -1) { 
    
-      const size_t register_offset = ichan % swtpg_wibeth::SAMPLES_PER_REGISTER;
-      const size_t register_t0_start = register_index * swtpg_wibeth::SAMPLES_PER_REGISTER * swtpg_wibeth::FRAMES_PER_MSG;
+      const size_t register_offset = ichan % swtpg_wib2::SAMPLES_PER_REGISTER;
+      const size_t register_t0_start = register_index * swtpg_wib2::SAMPLES_PER_REGISTER * swtpg_wib2::FRAMES_PER_MSG;
   
-      for (size_t iframe = 0; iframe<swtpg_wibeth::FRAMES_PER_MSG; ++iframe) {
+      for (size_t iframe = 0; iframe<swtpg_wib2::FRAMES_PER_MSG; ++iframe) {
     
-        const size_t msg_index = iframe / swtpg_wibeth::FRAMES_PER_MSG; 
-        const size_t msg_time_offset = iframe % swtpg_wibeth::FRAMES_PER_MSG;
+        const size_t msg_index = iframe / 12;
+        const size_t msg_time_offset = iframe % 12;
         // The index in uint16_t of the start of the message we want // NOLINT 
-        const size_t msg_start_index = msg_index * (swtpg_wibeth::ADCS_SIZE) / sizeof(uint16_t); // NOLINT
-        const size_t offset_within_msg = register_t0_start + swtpg_wibeth::SAMPLES_PER_REGISTER * msg_time_offset + register_offset;
+        const size_t msg_start_index = msg_index * swtpg_wib2::ADCS_SIZE / sizeof(uint16_t); // NOLINT
+        const size_t offset_within_msg = register_t0_start + swtpg_wib2::SAMPLES_PER_REGISTER * msg_time_offset + register_offset;
         const size_t index = msg_start_index + offset_within_msg;
     
         int16_t adc_value = input16[index];
-        std::cout << adc_value << std::endl;
+        //std::cout << adc_value << std::endl;
         out_file << ichan << "," <<  adc_value << "," << t_current << std::endl;
         t_current += 32;
       } 
@@ -166,23 +165,27 @@ void save_raw_data(swtpg_wibeth::MessageRegisters register_array,
 
 }
 
+
 // =================================================================
 //                       TPG FUNCTIONS
 // =================================================================
+
 void extract_hits_naive(uint16_t* output_location, uint64_t timestamp) {
 
     constexpr int clocksPerTPCTick = 32;
+    //uint16_t chan[100], hit_end[100], hit_charge[100], hit_tover[100]; 
     uint16_t chan, hit_end, hit_charge, hit_tover; 
+    unsigned int nhits = 0;
 
     size_t i = 0;
-    while (*output_location != swtpg_wibeth::MAGIC) {
+    while (*output_location != swtpg_wib2::MAGIC) {
       chan   = *output_location++;
       hit_end    = *output_location++;
       hit_charge  = *output_location++;
       hit_tover     = *output_location++;
 
 
-      //if (hit_charge && chan != swtpg_wibeth::MAGIC) {
+      //if (hit_charge && chan != swtpg_wib2::MAGIC) {
       //  std::cout << "Channel number: " << chan << std::endl;
       //  std::cout << "Hit charge: " << hit_charge << std::endl;
       //}
@@ -212,8 +215,12 @@ void extract_hits_naive(uint16_t* output_location, uint64_t timestamp) {
         save_hit_data(trigprim, "NAIVE");
       }
       ++total_hits;
+      
+      
 
     }
+
+
 }
 
 void extract_hits_avx(uint16_t* output_location, uint64_t timestamp) {
@@ -221,7 +228,7 @@ void extract_hits_avx(uint16_t* output_location, uint64_t timestamp) {
   constexpr int clocksPerTPCTick = 32;
   uint16_t chan[16], hit_end[16], hit_charge[16], hit_tover[16]; 
 
-  while (*output_location != swtpg_wibeth::MAGIC) {
+  while (*output_location != swtpg_wib2::MAGIC) {
     for (int i = 0; i < 16; ++i) {
       chan[i] = *output_location++; 
     }
@@ -240,9 +247,9 @@ void extract_hits_avx(uint16_t* output_location, uint64_t timestamp) {
     // find the channels which actually had a hit, as indicated by
     // nonzero value of hit_charge
     for (int i = 0; i < 16; ++i) {
-      if (hit_charge[i] && chan[i] != swtpg_wibeth::MAGIC) {
-        //std::cout << "Channel number: " << chan[i] << std::endl;
-        //std::cout << "Hit charge: " << hit_charge[i] << std::endl;
+      if (hit_charge[i] && chan[i] != swtpg_wib2::MAGIC) {
+        std::cout << "Channel number: " << chan[i] << std::endl;
+        std::cout << "Hit charge: " << hit_charge[i] << std::endl;
 
 
           uint64_t tp_t_begin =                                                        // NOLINT(build/unsigned)
@@ -256,6 +263,7 @@ void extract_hits_avx(uint16_t* output_location, uint64_t timestamp) {
           // For quick n' dirty debugging: print out time/channel of hits.
           // Can then make a text file suitable for numpy plotting with, eg:
           //
+          // sed -n -e 's/.*Hit: \(.*\) \(.*\).*/\1 \2/p' log.txt  > hits.txt
           //
           //TLOG_DEBUG(0) << "Hit: " << tp_t_begin << " " << offline_channel;
 
@@ -275,6 +283,8 @@ void extract_hits_avx(uint16_t* output_location, uint64_t timestamp) {
             save_hit_data(trigprim, "AVX");
           }          
 
+
+
         ++total_hits;
       }
     } // loop over 16 registers   
@@ -284,16 +294,15 @@ void extract_hits_avx(uint16_t* output_location, uint64_t timestamp) {
 
 
 
-void execute_tpg(const dunedaq::fdreadoutlibs::types::DUNEWIBEthTypeAdapter* fp) {
+void execute_tpg(const dunedaq::fdreadoutlibs::types::DUNEWIBSuperChunkTypeAdapter* fp) {
 
-  // Set CPU affinity of the TPG thread
   SetAffinityThread(0);
 
-  // Parse the WIBEth frames
-  auto wfptr = reinterpret_cast<dunedaq::fddetdataformats::WIBEthFrame*>((uint8_t*)fp);
+  // Parse the DUNEWIB frames
+  auto wfptr = reinterpret_cast<dunedaq::fddetdataformats::WIB2Frame*>((uint8_t*)fp);
   uint64_t timestamp = wfptr->get_timestamp();      
-  swtpg_wibeth::MessageRegisters registers_array;
-  swtpg_wibeth::expand_wibeth_adcs(fp, &registers_array);
+  swtpg_wib2::MessageRegisters registers_array;
+  swtpg_wib2::expand_wib2_adcs(fp, &registers_array, 0);
 
   
   if (first_hit) {                     
@@ -308,20 +317,27 @@ void execute_tpg(const dunedaq::fdreadoutlibs::types::DUNEWIBEthTypeAdapter* fp)
 
   }  
        
+  
   fh.m_tpg_processing_info->input = &registers_array;
   uint16_t* destination_ptr = fh.get_hits_dest();
-  *destination_ptr = swtpg_wibeth::MAGIC;
+  *destination_ptr = swtpg_wib2::MAGIC;
   fh.m_tpg_processing_info->output = destination_ptr;
-  m_assigned_tpg_algorithm_function(*fh.m_tpg_processing_info);
-       
+  //swtpg_wib2::process_window_avx2(*fh.m_tpg_processing_info, 0);       
+  m_assigned_tpg_algorithm_function(*fh.m_tpg_processing_info, 0);
+  //m_assigned_tpg_algorithm_function(*fh.m_tpg_processing_info, swtpg_wib2::NUM_REGISTERS_PER_FRAME * swtpg_wib2::SAMPLES_PER_REGISTER);
+          
+
+
   // Insert output of the AVX processing into the swtpg_output 
-  //swtpg_output swtpg_processing_result = {destination_ptr, timestamp};
+  swtpg_output swtpg_processing_result = {destination_ptr, timestamp};
     
   if (select_implementation == "AVX") {
     extract_hits_avx(destination_ptr, timestamp);
   } else if(select_implementation == "NAIVE") {  
     extract_hits_naive(destination_ptr, timestamp);
   }
+   
+
 
 }
 
@@ -329,6 +345,8 @@ void execute_tpg(const dunedaq::fdreadoutlibs::types::DUNEWIBEthTypeAdapter* fp)
 // =================================================================
 //                       MAIN
 // =================================================================
+
+
 int 
 main(int argc, char** argv)
 {
@@ -336,32 +354,32 @@ main(int argc, char** argv)
     CLI::App app{ "Test TPG algorithms" };
 
     // Set default input frame file
-    std::string frame_file_path = "./wibeth-frames.bin";
+    std::string frame_file_path = "./wib2-frames.bin";
     app.add_option("-f,--frame_file_path", frame_file_path, "Path to the input frame file");
 
     app.add_option("-a,--algorithm", select_algorithm, "TPG Algorithm (SimpleThreshold / AbsRS)");
   
     app.add_option("-i,--implementation", select_implementation, "TPG implementation (AVX / NAIVE)");
 
-    int num_frames_to_read = -1;
-    app.add_option("-n,--num_frames_to_read", num_frames_to_read, "Number of frames to read. Default: select all frames.");
+    int num_frames = -1;
+    app.add_option("-n,--num_frames", num_frames, "Number of frames to read. Default: select all frames.");
 
-    int swtpg_threshold = 500;
+    int swtpg_threshold = 100;
     app.add_option("-t,--swtpg_threshold", swtpg_threshold, "Value of the TPG threshold");
 
-    app.add_option("--save_adc_data", save_adc_data, "Save ADC data (true/false)");
+    app.add_flag("--save_adc_data", save_adc_data, "Save ADC data");
 
-    app.add_option("--save_trigprim", save_trigprim, "Save trigger primitive data (true/false)");
+    app.add_flag("--save_trigprim", save_trigprim, "Save trigger primitive data");
 
 
     CLI11_PARSE(app, argc, argv);
 
     if (select_algorithm == "SimpleThreshold") {
       if (select_implementation == "NAIVE") {
-        m_assigned_tpg_algorithm_function = &swtpg_wibeth::process_window_naive<swtpg_wibeth::NUM_REGISTERS_PER_FRAME>;
+        m_assigned_tpg_algorithm_function = &swtpg_wib2::process_window_naive<swtpg_wib2::NUM_REGISTERS_PER_FRAME>;
         std::cout << "Created an instance of the " << select_algorithm << " algorithm ( " << select_implementation << " )" << std::endl;
       } else if (select_implementation == "AVX") {
-        m_assigned_tpg_algorithm_function = &swtpg_wibeth::process_window_avx2<swtpg_wibeth::NUM_REGISTERS_PER_FRAME>;
+        m_assigned_tpg_algorithm_function = &swtpg_wib2::process_window_avx2<swtpg_wib2::NUM_REGISTERS_PER_FRAME>;
         std::cout << "Created an instance of the " << select_algorithm << " algorithm ( " << select_implementation << " )" << std::endl;
       } else {
         std::cout << "Select a valid algorithm implementation. Use --help for further details." << std::endl;
@@ -369,10 +387,10 @@ main(int argc, char** argv)
       }
     } else if (select_algorithm == "AbsRS") {
       if (select_implementation == "NAIVE") {        
-        m_assigned_tpg_algorithm_function = &swtpg_wibeth::process_window_naive_RS<swtpg_wibeth::NUM_REGISTERS_PER_FRAME>;
+        //m_assigned_tpg_algorithm_function = &swtpg_wib2::process_window_naive_RS<swtpg_wib2::NUM_REGISTERS_PER_FRAME>;
         std::cout << "Created an instance of the " << select_algorithm << " algorithm ( " << select_implementation << " )" << std::endl;
       } else if (select_implementation == "AVX") {
-        m_assigned_tpg_algorithm_function = &swtpg_wibeth::process_window_rs_avx2<swtpg_wibeth::NUM_REGISTERS_PER_FRAME>;
+        m_assigned_tpg_algorithm_function = &swtpg_wib2::process_window_rs_avx2<swtpg_wib2::NUM_REGISTERS_PER_FRAME>;
         std::cout << "Created an instance of the " << select_algorithm << " algorithm ( " << select_implementation << " )" << std::endl;
       } else {
         std::cout << "Select a valid algorithm implementation. Use --help for further details." << std::endl;
@@ -390,79 +408,77 @@ main(int argc, char** argv)
     // =================================================================
     const std::string input_file = frame_file_path;
     std::unique_ptr<dunedaq::readoutlibs::FileSourceBuffer> m_source_buffer;
-    m_source_buffer = std::make_unique<dunedaq::readoutlibs::FileSourceBuffer>(10485100, WIBEth_FRAME_SIZE); // AAA: hardcoded values!! 
+    m_source_buffer = std::make_unique<dunedaq::readoutlibs::FileSourceBuffer>(10485100, swtpg_wib2::SUPERCHUNK_FRAME_SIZE);
  
     m_source_buffer->read(input_file);
     auto& source = m_source_buffer->get(); 
-    int total_num_frames = m_source_buffer->num_elements(); // file_ size/chunk_size = 180 
+    int num_superchunks = m_source_buffer->num_elements(); // file_ size/chunk_size = 180 
 
-    std::cout << "Number of DUNE WIBEth frames in the input file: " << total_num_frames << std::endl;
+    std::cout << "Number of superchunks in the input file: " << num_superchunks << std::endl;
 
     // Check if the selected number of frames is <= than the ones available in the input file
-    if (total_num_frames < num_frames_to_read) {
+    if (num_superchunks < num_frames) {
       std::cout << "\n**ERROR**: Select a valid number of frames that is less or equal to the ones available in the input file." << std::endl;
       return 1;
-    } else if (num_frames_to_read == -1) {
-      num_frames_to_read = total_num_frames;
+    } else if (num_frames == -1) {
+      num_frames = num_superchunks;
     }
 
-    
+
     // =================================================================
     //                       Setup the SWTPG
     // =================================================================
 
-    auto limiter = dunedaq::readoutlibs::RateLimiter(31);
+    auto limiter = dunedaq::readoutlibs::RateLimiter(166);
     limiter.init();
 
     fh.initialize(swtpg_threshold);
 
     // =================================================================
-    //                  Process the DUNE WIBEth frames
+    //                       Process the DUNEWIB superchunks
     // =================================================================
-    int wibeth_frame_index = 0; 
-    uint64_t frame_repeat_index = 0;
+    int superchunk_index = 0; 
+    uint64_t frame_repeat = 0;
     auto start_test = std::chrono::high_resolution_clock::now();  
 
-    // Loop over the DUNEWIB Ethernet frames in the file
-    while (wibeth_frame_index < num_frames_to_read ){      
-
-      // current WIBEth frame
-      auto fp = reinterpret_cast<dunedaq::fdreadoutlibs::types::DUNEWIBEthTypeAdapter*>(source.data() + wibeth_frame_index*WIBEth_FRAME_SIZE);
+    // Loop over the DUNEWIB superchunks in the file
+    while (superchunk_index < num_frames ){      
+    //while (true){        
+      
+      // current superchunk
+      auto fp = reinterpret_cast<dunedaq::fdreadoutlibs::types::DUNEWIBSuperChunkTypeAdapter*>(source.data() + superchunk_index*swtpg_wib2::SUPERCHUNK_FRAME_SIZE);
 
       execute_tpg(fp);
 
-      ++wibeth_frame_index;
+      ++superchunk_index;
+      //if (superchunk_index % 10 == 0) {
+      //  std::cout << "Executing superchunk number " << superchunk_index << " out of " << num_superchunks << std::endl;        
+      //}
 
-
-      // If end of the file is reached, restart the index counter
-      if (wibeth_frame_index == num_frames_to_read) {
-        wibeth_frame_index = 0;
-	frame_repeat_index++;
+      // If end of the file is reached, measure the elapsed time
+      if (superchunk_index == num_frames) {
+        superchunk_index = 0;
+	frame_repeat++;
       }
-
-      if (frame_repeat_index % 500  == 0) {
+      if (frame_repeat % 500 == 0) {
         // Calculate elapsed time in seconds  
         auto now = std::chrono::high_resolution_clock::now();
         auto elapsed_seconds = std::chrono::duration_cast<std::chrono::seconds>(now - start_test).count();  
-        std::cout << "Elapsed time [s]: " << elapsed_seconds << std::endl;      
-	
-	frame_repeat_index = 0;        
-
-        // stop the testing after a time a condition
-        if (elapsed_seconds > TIME_TEST) {
-          wibeth_frame_index = num_frames_to_read;
-        }
+        std::cout << "Elapsed time [s]: " << elapsed_seconds << std::endl;
+	frame_repeat = 0;
       }
-
 	
 
       limiter.limit();
 
     }
+
     std::cout << "\n\n===============================" << std::endl;
     std::cout << "Found in total " << total_hits << " hits." << std::endl;
     
     std::cout << "\n\nFinished testing." << std::endl;
+
+
 
 }
